+++ extension-manifest-v2/CHANGELOG.md
@@ -1,11 +1,12 @@
 ### GHOSTERY 8.4.4 (UNRELEASED)
 
-+ New upgrade modals for Ghostery Plus & Ghostery Insights (#464)
++ New upgrade modals for Ghostery Plus & Ghostery Insights (#458,#460,#463,#464,#467,#468)
 + Improves context-detection of WebRequests, particularly for pages using service workers (#461)
 + Fixes some cases where cookies were blocked for first parties because anti-tracking thought they were 3rd. (#461)
 + Updates Ad Blocker version. We now support HTML filters on Firefox. (#461)
 + Human Web cleanup and refactor (#461)
-+ Updated telemetry ()
++ Updated telemetry (#466)
++ Updated translations
 + Language string cleanup
 
 ### GHOSTERY 8.4.3 (October 1, 2019)